#IUFC-384 Ajout d'attributs dans le fichier xls produit pour les inscriptions/archives
