+++ business/xls/xls_archive.py
@@ -28,6 +28,8 @@ from django.utils.translation import ugettext_lazy as _
 from base.business.xls import get_name_or_username
 from osis_common.document import xls_build
 from continuing_education.business.xls.xls_common import form_filters
+from continuing_education.business.xls.xls_common import get_titles_admission, get_titles_registration, \
+    extract_xls_data_from_registration
 
 ARCHIVE_TITLES = [
     str(_('Name')),
@@ -50,7 +52,7 @@ def create_xls(user, archive_list, form):
     parameters = {xls_build.DESCRIPTION: XLS_DESCRIPTION,
                   xls_build.USER: get_name_or_username(user),
                   xls_build.FILENAME: XLS_FILENAME,
-                  xls_build.HEADER_TITLES: ARCHIVE_TITLES,
+                  xls_build.HEADER_TITLES: _get_titles(),
                   xls_build.WS_TITLE: WORKSHEET_TITLE}
 
     return xls_build.generate_xls(xls_build.prepare_xls_parameters_list(working_sheets_data, parameters), filters)
@@ -61,11 +63,9 @@ def prepare_xls_content(archive_list):
 
 
 def extract_xls_data_from_archive(archive):
-    return [
-        archive.person_information.person.last_name,
-        archive.person_information.person.first_name,
-        archive.email,
-        archive.formation_display,
-        archive.get_faculty() if archive.get_faculty() else '',
-        _(archive.state) if archive.state else ''
-    ]
+    return extract_xls_data_from_registration(archive)
+
+
+def _get_titles():
+    return get_titles_admission() + get_titles_registration()
+
