+++ tests/business/xls/__init__.py
@@ -1,0 +1,1 @@
+__author__ = 'verpoorten'
